account: Separate concerns
